[Style/#51] Separate HomeView from MainView
MainView가 길어서 코드분리를 하였다
